Fix resolution failure not propagated to build operation
This commit fixes the resolution result errors not being set on the
build operation type, which was a regression from 4.8. The reason is
that failures are thrown lazily, when the resolution result is used,
rather than eagerly. However, we *do* know if consuming the result
would fail, we just have to capture that information before.